Fix wrapping for lines composed only of indentation
This commit fixes a bug in the `wrapLine()` algorithm which caused a single line comprised only of indentation to be printed before any other lines, in the case where the the input's first word was too long for the allowed column length.

This could be observed by going to https://domenic.github.io/rewrapper and inputting:

```
       <p>This-specification-is-intended-for-authors-of-documents-and-scripts-that-use-the-features-blah-blah-blah.
```

The reason the output would have a random preceding newline full of only indentation (but only 6 spaces, instead of 7!), is because each space in the indentation was being considered as a word of length 0. This meant that for each space in the initial indentation, we'd hit the first path of the `if` condition in the `for` loop, which would incrementally build `currentLine` up to to `leadingIndent.length` - ` number of `spaceBefore`s.

Once we finally got to the first real word, if that word made `currentLine` too long, we would commit `currentLine` (of only indentation) to `brokenLines` and continue.

This is wrong and weird. To prevent us from committing a line of only spaces to the output, this commit adds a check to ensure that we don't do this.

That leaves us with the fact that we're still computing the first line really weird, by treating each leading space as a separate word. For all subsequent lines, we eagerly tack on `leadingIndent` to the beginning of `currentLine` when `currentLine` is reset, so we should just do the same for the first line. This commit changes the initialization of `currentLine` to equal `leadingIndent` initially, and invokes `line.trimLeft()` right away in the function.